requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
tensorflow==2.15.0
MetaTrader5==5.0.37
python-binance==1.0.15
nltk==3.6.3
newsapi-python==0.2.6
plotly==5.3.1
dash==2.0.0
joblib==1.1.0
scikit-learn==1.0.2
